consolidate target vetting into hooks.py
